pole-app/backend/requirements.txt
flask
flask-sqlalchemy
flask-bcrypt
python-dotenv
flask-session
pyjwt